Added more functionality to boggle_board
Board now prints a clock to the system, and can take parameters for players and difficulty. We need to get a board state in here to place in the squares, add a text entry box, and add a text print box. I don't know if this screen will wind up replacing main.py or not, so we need to decide on that